refactor(chat): extract prompt_builder.py (classification, reader context, LLM messages)
Verbatim move of SYSTEM_PROMPT / META_INTRO_PROMPT / _META_KEYWORDS /
_FAILED_ANSWER_PREFIXES and _is_meta_question, _classify_and_enhance_prompt,
_build_reader_context_prompt, _build_reader_data_block, _estimate_tokens,
_build_llm_messages, _strip_followup_suggestions, _is_failed_answer.
Pure functions — no DB/HTTP/streaming. Re-imported into app.services.chat
(eval/run_eval.py imports _build_llm_messages from there). 819 tests pass.
